+++ src/eventsub/channel/subscription/message.rs
@@ -111,7 +111,7 @@ pub mod tests {
             }
             "##;
 
-        let val = dbg!(crate::eventsub::Payload::parse(payload).unwrap());
+        let val = dbg!(crate::eventsub::Event::parse(payload).unwrap());
         crate::tests::roundtrip(&val)
     }
     #[test]