+++ src/helix/subscriptions/get_broadcaster_subscriptions_events.rs
@@ -14,9 +14,10 @@
 //!
 //! ```rust
 //! use twitch_api2::helix::subscriptions::get_broadcaster_subscriptions_events;
-//! let request = get_broadcaster_subscriptions_events::GetBroadcasterSubscriptionsEventsRequest::builder()
-//!     .broadcaster_id("1234")
-//!     .build();
+//! let request =
+//!     get_broadcaster_subscriptions_events::GetBroadcasterSubscriptionsEventsRequest::builder()
+//!         .broadcaster_id("1234")
+//!         .build();
 //! ```
 //!
 //! ## Response: [BroadcasterSubscriptionEvent]